requirements.txt
requests
pillow
diffusers
transformers
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
opencv-python
datetime
flask
openai
boto3
python-dotenv
celery
redis
flake8
